start button register user in db

+++ database/models.py
@@ -1,11 +1,12 @@
 from database.engine import Base
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Integer, String, BigInteger
+
 
 class User(Base):
     __tablename__ = "users"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False, index=True)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False, index=True)
     username: Mapped[str | None] = mapped_column(String(50), nullable=True)
     full_name: Mapped[str | None] = mapped_column(String(100), nullable=True)
